Remove RtMidi and Garry's Mod files
Headers and source code for RtMidi and Garry's Mod are no longer
included with the project to avoid forcing specific versions and having
to update them when the upstream releases new versions. It will be up
to the person compiling the module to get these files.

This also means that RtMidi is no longer compiled by this project,
so the library will need to be obtained some other way. Many Linux
distributions provide the headers, source and compiled libraries for
RtMidi in their official package repositories, which is recommended
if possible.

If a manually compiled version of RtMidi is still desired, Premake will
look for it in a lib32 or lib64 directory in the root directory of the
repository. Of course, a static library is always recommended.

I might further change how all this works, or document it better, but
for now this is how things will be.