Add RtMidi and Garry's Mod source/headers
Decided to revert the removal of source code and headers for RtMidi
and Garry's Mod due to the custom nature in which RtMidi is built,
and the Garry's Mod headers are just annoying to get manually.

The source code for gmcl_rtmidi and the Premake script were refactored
to be more compact with less whitespace.

On Windows, RtMidi is built with RTMIDI_DO_NOT_ENSURE_UNIQUE_PORTNAMES,
which prevents the behavior of prepending the port ID to the port name,
making the behavior on Windows consistent with Linux and macOS.

The JACK backend option for Premake was re-added. The default backends
are selected based on the operating system, but if --jack is passed
to Premake then JACK will also be included, which takes priority over
the default backend if RtMidi determines it is available.

Some extra definitions are expliticly defined with Premake now. I'm not
really sure if they're defined automatically via the IDE/compiler, but
I figured it doesn't hurt to define them expliticly anyway.

The license for RtMidi was added in its own LICENSE file, which is also
mentioned in README.md.